+++ examples/e-commerce/src/search.ts
@@ -1,0 +1,53 @@
+import algoliasearch from 'algoliasearch/lite';
+import instantsearch from 'instantsearch.js';
+import {
+  brands,
+  categories,
+  clearFilters,
+  clearFiltersEmptyResults,
+  clearFiltersMobile,
+  configuration,
+  freeShipping,
+  hitsPerPage,
+  pagination,
+  priceSlider,
+  products,
+  ratings,
+  resultsNumberMobile,
+  saveFiltersMobile,
+  searchBox,
+  sortBy,
+} from './widgets';
+import routing from './routing';
+
+const searchClient = algoliasearch(
+  'latency',
+  '6be0576ff61c053d5f9a3225e2a90f76'
+);
+
+const search = instantsearch({
+  searchClient,
+  indexName: 'instant_search',
+  routing,
+});
+
+search.addWidgets([
+  brands,
+  categories,
+  clearFilters,
+  clearFiltersEmptyResults,
+  clearFiltersMobile,
+  configuration,
+  freeShipping,
+  hitsPerPage,
+  pagination,
+  priceSlider,
+  products,
+  ratings,
+  resultsNumberMobile,
+  saveFiltersMobile,
+  searchBox,
+  sortBy,
+]);
+
+export default search;